+++ cypress/plugins/index.js
@@ -41,6 +41,7 @@ const tempReportEnvJson = '../../tempReportEnv.json';
 const { getAndDereferenceOpenRpc } = require('./pluginUtils');
 let metaDataArr = [];
 let maxExitCode = 0;
+const { fireLog } = require('../support/cypress-support/src/fireLog');
 
 module.exports = async (on, config) => {
   // To set the specPattern dynamically based on the testSuite
@@ -232,7 +233,7 @@ module.exports = async (on, config) => {
   });
 
   on('before:run', async () => {
-    console.debug('Entering before:run in cypress/plugins/index.js');
+    fireLog.info('Entering before:run in cypress/plugins/index.js');
 
     // Calling reportProcessor default function with instance of event
     const reportProcessor = importReportProcessor();
@@ -256,7 +257,7 @@ module.exports = async (on, config) => {
       - generate the html report (TBD)
   */
   on('after:run', async (results) => {
-    console.debug('Entering after :run in cypress/plugins/index.js');
+    fireLog.info('Entering after :run in cypress/plugins/index.js');
 
     const reportObj = {};
     const formatter = new Formatter();
@@ -289,9 +290,9 @@ module.exports = async (on, config) => {
       if (!fs.existsSync(filePath)) {
         try {
           fs.mkdirSync(filePath);
-          console.debug('Cucumber-json folder created successfully.');
+          fireLog.info('Cucumber-json folder created successfully.');
         } catch (error) {
-          console.error(error);
+          fireLog.info(error);
         }
       }
 
@@ -308,7 +309,7 @@ module.exports = async (on, config) => {
       // delete the messages.ndjson file.
       fs.unlink(sourceFile, (err) => {
         if (err) throw err;
-        console.debug('The file has been deleted!');
+        fireLog.info('The file has been deleted!');
       });
 
       const reportType = config.env.reportType;
@@ -339,20 +340,21 @@ module.exports = async (on, config) => {
               // to remove the wait time step from html report
               const bufferString = jsonReport.toString();
               const parsedJson = JSON.parse(bufferString);
-              parsedJson.forEach((obj) => {
-                obj.elements = obj.elements.map((element) => {
-                  const filteredSteps = [];
-                  element.steps.forEach((step, index) => {
-                    if (/^Test runner waits for/.test(step.name)) {
-                      console.log(`Removing step ${index} from html report:`, step.name);
-                    } else {
-                      filteredSteps.push(step);
-                    }
+              Array.isArray(parsedJson) &&
+                parsedJson.forEach((obj) => {
+                  obj.elements = obj.elements.map((element) => {
+                    const filteredSteps = [];
+                    element.steps.forEach((step, index) => {
+                      if (/^Test runner waits for/.test(step.name)) {
+                        fireLog.info(`Removing step ${index} from html report:`, step.name);
+                      } else {
+                        filteredSteps.push(step);
+                      }
+                    });
+                    element.steps = filteredSteps;
+                    return element;
                   });
-                  element.steps = filteredSteps;
-                  return element;
                 });
-              });
               const updatedJsonReport = JSON.stringify(parsedJson, null, 2);
               jsonReport = Buffer.from(updatedJsonReport);
             }
@@ -407,11 +409,11 @@ module.exports = async (on, config) => {
       }
       process.exitCode = maxExitCode;
     } catch (err) {
-      console.error('Error occurred in after:run hook:' + err);
+      fireLog.info('Error occurred in after:run hook:' + err);
       process.exitCode = CONSTANTS.FCS_EXIT_CODE.FAILURE;
     }
 
-    console.log(`FCS_EXIT_CODE: ${process.exitCode}`);
+    fireLog.info(`FCS_EXIT_CODE: ${process.exitCode}`);
   });
   return config;
 };