+++ cypress/plugins/pluginUtils.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 const path = require('path');
 const $RefParser = require('@apidevtools/json-schema-ref-parser');
 const _ = require('lodash');
+const { fireLog } = require('../support/cypress-support/src/fireLog');
 
 /**
  * Fetches and dereferences OpenRPC documents from various sources including a Firebolt URL, local files, and external URLs.
@@ -62,7 +63,7 @@ async function getAndDereferenceOpenRpc(externalUrls, version = null, sdkSignOff
     }
     return openRpcDocs;
   } catch (err) {
-    console.error(err, 'getAndDereferenceOpenRpc');
+    fireLog.info(err, 'getAndDereferenceOpenRpc');
   }
 }
 
@@ -117,7 +118,7 @@ function generateIndexFile(path, outputObj) {
     // Write to the new index.js file
     fs.writeFileSync(indexFilePath, indexFileContent);
   } catch (error) {
-    console.error(`An error occurred while generating the index file: ${error}`);
+    fireLog.info(`An error occurred while generating the index file: ${error}`);
     throw error;
   }
 }
@@ -134,7 +135,7 @@ function preprocessDeviceData(config) {
   const deviceMac = config.env.deviceMac;
   try {
     if (!deviceMac) {
-      console.error('Device MAC address is required.');
+      fireLog.info('Device MAC address is required.');
     }
     const formattedDeviceMac = deviceMac.replace(/:/g, '').toUpperCase();
     const jsonFilePath = `cypress/fixtures/external/devices/${formattedDeviceMac}.json`;
@@ -143,7 +144,7 @@ function preprocessDeviceData(config) {
     try {
       deviceData = JSON.parse(fs.readFileSync(jsonFilePath, 'utf-8'));
     } catch (readError) {
-      console.error(
+      fireLog.info(
         `Error reading or parsing the JSON file at ${jsonFilePath}: ${readError.message}`
       );
     }
@@ -151,7 +152,7 @@ function preprocessDeviceData(config) {
     const resolvedDeviceData = { ...deviceData };
     config.env = Object.assign({}, config.env, { resolvedDeviceData });
   } catch (error) {
-    console.error(`Error in preprocessDeviceData: ${error.message}`);
+    fireLog.info(`Error in preprocessDeviceData: ${error.message}`);
   }
 }